Add Coinbase live feed for BTC-USD and ETH-USD
- Add python/data/coinbase_feed.py subscribing to Coinbase level2 channel
- Add python/data/live_runner.py with persistent per-product OrderBook state
- Fix quantity type from int to double across Event, Order, Position structs
- Update mean_reversion.py parameters for live crypto data
- Update .gitignore to exclude data/*.log and .env
- Update README with live results: BTC 42 trades +\$0.9114, ETH 30 trades +\$0.0273

+++ src/engine/event.hpp
@@ -11,6 +11,6 @@ struct Event {
     EventType type;
     int       order_id;
     double    price;
-    int       quantity;
+    double    quantity;   // changed from int to double
     bool      is_buy;
 };
